Move classes other than Experiment subclass to separate module
